chore(cli): Moved MCP.ts to mcp directory, added mcp/README.md, and removed platform CLI commands from the MCP server
